pyproject.toml
[project]
name = "automatic-transmission"
version = "5.7.14"
description = "Media automation system for RSS feeds and Plex"
requires-python = ">=3.12"

dependencies = [
    "dbt-postgres>=1.9.0",
    "feedparser>=6.0.11",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pandera[polars]>=0.20.0",
    "paramiko>=3.5.1",
    "polars[timezone]>=1.29.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "transmission-rpc>=7.0.11",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "tests*"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]